# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * WerkParts - Collision Fastener Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from './categoryCount';
import type { Invoice } from './invoice';

export interface DashboardStats {
  totalInvoices: number;
  totalRevenue: string;
  invoicesThisMonth: number;
  revenueThisMonth: string;
  recentInvoices: Invoice[];
  partsByCategory: CategoryCount[];
}