# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * WerkParts - Collision Fastener Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Invoice {
  id: number;
  invoiceNumber: string;
  roNumber: string;
  date: string;
  techName: string;
  vehicleYear: string;
  vehicleMake: string;
  vehicleModel: string;
  insuranceCompany: string;
  /** draft | finalized | paid */
  status: string;
  totalAmount: string;
  itemCount?: number;
  createdAt: string;
}